Updated code and refactor
Refactored some of the code to reduce the use of event listeners.

added functionality to the delete button. finds the index of the array and removes the original with array with splice.

added some checks to prevent NaN error when submit is pressed without any input.

added unique ID to each element to aid in removing from array and preventing accidental deletions.